% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_sp_metadata_v3}
\alias{Rcpp_sp_metadata_v3}
\title{Get metadata from a dense SPZ file}
\usage{
Rcpp_sp_metadata_v3(path)
}
\arguments{
\item{path}{Input file path .spz (dense)}
}
\value{
A list with file metadata
}
\description{
Get metadata from a dense SPZ file
}
\keyword{internal}
