% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_sp_read_dense}
\alias{Rcpp_sp_read_dense}
\title{Read a dense SPZ file into a numeric matrix}
\usage{
Rcpp_sp_read_dense(path)
}
\arguments{
\item{path}{Input file path .spz (dense)}
}
\value{
A numeric matrix
}
\description{
Read a dense SPZ file into a numeric matrix
}
\keyword{internal}
